+++ client/src/components/Profile/DetailPage.js
