+++ client/src/components/Profile/DetailPage.scss
