+++ client/src/components/Profile/components/DetailAssets.js
