+++ client/src/components/Profile/components/DetailAssets.scss
