+++ client/src/components/Profile/components/Politician.js
